Doc: Switch to doxygen, breathe and sphinx for documentation generation
Doxygen limitations:
- Markdown support does not suffice for our documentation needs
- No fulltext search available
- No syntax highlightning of code blocks for Igor Pro

Therefore we now use only the XML output of doxygen and render that to
html using breathe and sphinx.

As we still have some documentation in markdown we want to keep we convert
these to rst on the fly using pandoc.

The following previously supported features of doxyen don't work anymore:
- Todo/deprecated lists
- call/callee graph support

For local testing purposes the pure doxygen html output can still be
generated.