IPNWB: Avoid error on saving waves with very large wave notes
Since forever we store the wave note with the wave data in the NWB file.
By accident one of the MIES users encountered a sweep wave with a very
large size, due to ED_WriteChangedValuesToNote and
ED_WriteChangedValuesToNoteText, and this triggerd an error on calling
HDF5SaveData.

This error is a HDF5 limitation as by default attributes are stored in the
object header and this header is limited to a total size of 64k.

In principle the HDF5 XOP could use "dense attributes" introduced with
HDF5 1.8.0.

We work around this limitation by truncating the wave note if required.