CreateIntraCellularEphys: Overwrite by default
Since 437674a9 (NWB Export: Only create intracelluar ephys group if writing
sweep data, 2017-03-02) we make the call to CreateIntraCellularEphys only
if we are saving sweep data.

We habe two call sites for CreateIntraCellularEphys. One is
NWB_ExportAllData and the other NWB_AppendSweep. The call in NWB_AppendSweep is
required as we might add sweep data to an NWB file without prior sweep
data. But as we call NWB_AppendSweep for every sweep we have to overwrite
the dataset written in CreateIntraCellularEphys.