Rename chunkedLayout parameter and add single chunk compression mode
Ever since the introduction of chunkedLayout setting this to 1 did two
things. It used chunking/shuffling *and* enabled GZIP compression.

As the size of the chunks is very fragile parameter to get right we
introduce a new mode, SINGLE_CHUNK_COMPRESSION, which just uses one
chunk for a dataset.

In order to properly test we need to convert all users of chunkedLayout
to the new compression mode flags.