Update Specifications for 1.0.5
The NWB schema for version 1.0.5 is updated to match up explicit `yaml`
syntax (indentless lists). The specifications are only loosely defined
and mostly resembles the structure of the corresponding HDF5 files.
NWB version 1 is a legacy (archived) file format description. For newer
implementations, please use NWB file format in a higher version (>=2)
where the specifications were hardened and the format has a wider
acceptance in the community.