Change H5_OpenGroup and H5_GroupExists Behaviour
In HDF5 XOP < 2.03 a link to a group cannot be opened using
`HDF5OpenGroup`. However, the XOP function was built deep into the
`H5_GroupExists` wrapper.

These changes separate the two functionaliteis of checking if a group
exists and opening a group. It will print to the history if a possibly
undefined group opening was encountered. The derived functions that
depend on `H5_GroupExists` (like `H5_AttributeExists` seem to handle
links in paths well, so it is only the opening of a group that causes
problems.

The introduced changes also allow testing attributes of linked groups
without failing at `H5_GroupExists`.

Note: The new version of the HDF5 XOP is shipped with IP 8.05.