H5_GroupExists: Speed it up
Since the bugfix in be9fb354 (Change H5_OpenGroup and H5_GroupExists
Behaviour, 2020-01-22) H5_GroupExists is painfully slow. The reason for
that fix was that HDF5OpenGroup did not work with soft links in its
path.

This is fixed in the (unreleased) HDF5 XOP 2.0.3 and in the current IP9
beta version.

So let's use HDF5OpenGroup for IP9 and HDF5ListAttributes for IP8. The
latter is deemed less expensive compared to the existing HDF5ListGroup.