Make using IPNWB in an independent module optional
In MIES we want to use the async framework to write NWB files using
IPNWB. This requires to call a threadsafe IPNWB function from a
threadsafe function in ProcGlobal.

This is not really supported in IP as it is buggy.

So let's make using the IM in IPNWB optional.

This also requires to allow skipping the builtin utility functions as we
have defined them in MIES already.