InitVectorIndex: Accept a VectorData
The VectorIndex class already contains a VectorData member, so let's
accept one in our init function as well. This has the advantage that we
can use the same pattern when writing the destination.

While at it we also rename timeseries_index to timeseries_vector_index as
we just append `_index` to the VectorData name in all other cases.