CreateIntraCellularEphys: Drop filtering argument
This is deprecated for pynwb 4.0 and dandi validate warns about via:

$ dandi validate --ignore "(NWBI|DANDI)" HardwareBasic-V2.nwb
C:\Users\Admin\.venvs\pynwb\Lib\site-packages\pynwb\core.py:52: UserWarning: Use of icephys_filtering is deprecated and will be removed in PyNWB 4.0. Use the IntracellularElectrode.filtering field instead

As we've never used the filtering, we are also just dropping it instead of
moving it to the IntracellularElectrode.